reconfigure bigquery client
