+++ src/evm-block-extractor/Cargo.toml
@@ -34,6 +34,7 @@ yup-oauth2 = { workspace = true }
 did = { path = "../did" }
 
 [dev-dependencies]
+rand = { workspace = true }
 tempfile = { workspace = true }
 testcontainers = { workspace = true }
 wiremock = { workspace = true }
