+++ src/evm-block-extractor/src/bin/server.rs
@@ -6,6 +6,11 @@ use jsonrpsee::RpcModule;
 
 #[derive(Debug, Clone, Parser)]
 pub struct ServerConfig {
+
+    /// The project ID of the BigQuery table
+    #[arg(long = "project-id", short('p'), default_value = "bitfinity-evm")]
+    project_id: String,
+
     /// The dataset ID of the BigQuery table
     /// The dataset ID can be one of the following:
     /// - `testnet`
@@ -42,7 +47,7 @@ async fn main() -> anyhow::Result<()> {
 
     let server = Server::builder().build(args.server_address).await?;
 
-    let db = BigQueryBlockChain::new(args.dataset_id, args.sa_key).await?;
+    let db = BigQueryBlockChain::new(args.project_id ,args.dataset_id, args.sa_key).await?;
 
     let eth = EthImpl::new(db);
