+++ src/evm-block-extractor/src/constants.rs
@@ -1,5 +1,3 @@
-pub const PROJECT_ID: &str = "bitfinity-evm";
-
 pub const RECEIPTS_TABLE_ID: &str = "receipts";
 
 pub const CHUNK_SIZE: usize = 10_000;
